VR: split GraphBuilder
